requirements.txt
daphne==4.0.0
dj-database-url==0.5.0
django-admin-interface==0.23.0
django-allauth==0.52.0
django-auditlog==2.2.2
django-axes==5.41.1
django-compressor==4.0
django-cors-headers==3.14.0
django-csvimport==2.17
django-debug-toolbar==4.0.0
django-extensions==3.2.0
django-hijack==3.3.0
django-jazzmin==2.6.0
django-rest-knox==4.2.0
django-treenode==0.19.0
drf-jwt==1.19.2
hashids==1.3.1
humanize==4.6.0
importlib-metadata==2.0.0
jsonfield==3.1.0
lazy-object-proxy==1.7.1
more-itertools==8.14.0
pip-chill==1.0.3
psycopg2==2.9.3
pyopenssl==23.1.1
python-decouple==3.6
regex==2022.10.31
reportlab==3.6.13
service-identity==21.1.0
simplejson==3.19.1
sqlalchemy==2.0.12
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.1
whitenoise==6.4.0